reorganized api transformers and corresponding tests
