+++ pliers/converters/__init__.py
@@ -3,13 +3,14 @@ of one type as input and return a `Stim` of a different type as output.
 '''
 
 from .api import (WitTranscriptionConverter,
-                  IBMSpeechAPIConverter)
+                  IBMSpeechAPIConverter,
+                  GoogleSpeechAPIConverter,
+                  GoogleVisionAPITextConverter,
+                  MicrosoftAPITextConverter)
 from .base import Converter, get_converter
-from .google import GoogleVisionAPITextConverter, GoogleSpeechAPIConverter
 from .image import TesseractConverter
 from .iterators import (VideoFrameIterator, VideoFrameCollectionIterator,
                         ComplexTextIterator)
-from .microsoft import MicrosoftAPITextConverter
 from .multistep import VideoToTextConverter, VideoToComplexTextConverter
 from .video import VideoToAudioConverter
