+++ pliers/converters/api/__init__.py
@@ -1,0 +1,12 @@
+from .wit import WitTranscriptionConverter
+from .ibm import IBMSpeechAPIConverter
+from .google import GoogleVisionAPITextConverter, GoogleSpeechAPIConverter
+from .microsoft import MicrosoftAPITextConverter
+
+__all__ = [
+    'WitTranscriptionConverter',
+    'IBMSpeechAPIConverter',
+    'GoogleSpeechAPIConverter',
+    'GoogleVisionAPITextConverter',
+    'MicrosoftAPITextConverter'
+]
