+++ pliers/extractors/api/__init__.py
@@ -1,0 +1,35 @@
+from .clarifai import ClarifaiAPIExtractor
+from .indico import (IndicoAPITextExtractor,
+                     IndicoAPIImageExtractor)
+from .google import (GoogleVisionAPIFaceExtractor,
+                     GoogleVisionAPILabelExtractor,
+                     GoogleVisionAPIPropertyExtractor,
+                     GoogleVisionAPISafeSearchExtractor,
+                     GoogleVisionAPIWebEntitiesExtractor)
+from .microsoft import (MicrosoftAPIFaceExtractor,
+                        MicrosoftAPIFaceEmotionExtractor,
+                        MicrosoftVisionAPIExtractor,
+                        MicrosoftVisionAPITagExtractor,
+                        MicrosoftVisionAPICategoryExtractor,
+                        MicrosoftVisionAPIImageTypeExtractor,
+                        MicrosoftVisionAPIColorExtractor,
+                        MicrosoftVisionAPIAdultExtractor)
+
+__all__ = [
+    'ClarifaiAPIExtractor',
+    'IndicoAPITextExtractor',
+    'IndicoAPIImageExtractor',
+    'GoogleVisionAPIFaceExtractor',
+    'GoogleVisionAPILabelExtractor',
+    'GoogleVisionAPIPropertyExtractor',
+    'GoogleVisionAPISafeSearchExtractor',
+    'GoogleVisionAPIWebEntitiesExtractor',
+    'MicrosoftAPIFaceExtractor',
+    'MicrosoftAPIFaceEmotionExtractor',
+    'MicrosoftVisionAPIExtractor',
+    'MicrosoftVisionAPITagExtractor',
+    'MicrosoftVisionAPICategoryExtractor',
+    'MicrosoftVisionAPIImageTypeExtractor',
+    'MicrosoftVisionAPIColorExtractor',
+    'MicrosoftVisionAPIAdultExtractor'
+]
