+++ pliers/tests/extractors/api/test_google_extractors.py
@@ -6,12 +6,12 @@ from pliers.extractors import (GoogleVisionAPIFaceExtractor,
                                GoogleVisionAPIWebEntitiesExtractor,
                                ExtractorResult,
                                merge_results)
-from pliers.extractors.google import GoogleVisionAPIExtractor
+from pliers.extractors.api.google import GoogleVisionAPIExtractor
 from pliers.stimuli import ImageStim, VideoStim
 import pytest
 import json
 from os.path import join
-from ..utils import get_test_data_path
+from ...utils import get_test_data_path
 import numpy as np
