+++ pliers/transformers/__init__.py
@@ -4,8 +4,10 @@ that have multiple subclasses spanning different modules (e.g., Google Cloud
 extractors that span audio, image, etc.).'''
 
 from .base import Transformer, BatchTransformerMixin, get_transformer
-from .google import GoogleAPITransformer, GoogleVisionAPITransformer
-from .microsoft import MicrosoftAPITransformer, MicrosoftVisionAPITransformer
+from .api import (GoogleAPITransformer,
+                  GoogleVisionAPITransformer,
+                  MicrosoftAPITransformer,
+                  MicrosoftVisionAPITransformer)
 
 
 __all__ = [
