+++ pliers/transformers/api/__init__.py
@@ -1,0 +1,9 @@
+from .google import GoogleAPITransformer, GoogleVisionAPITransformer
+from .microsoft import MicrosoftAPITransformer, MicrosoftVisionAPITransformer
+
+__all__ = [
+    'GoogleAPITransformer',
+    'GoogleVisionAPITransformer',
+    'MicrosoftAPITransformer',
+    'MicrosoftVisionAPITransformer'
+]
